Add options for ignoring line numbers and exception values in tracebacks.
Ignoring line numbers in tracebacks allows for easier processing of log
files where there may have been unrelated code changes in files that are
involved in a traceback. With --ignore-line-numbers, those tracebacks
will be grouped together when --stats is used.

Sometimes the exception VALUE is all that's different and so, if that
doesn't matter for the purposes of analysis, --ignore-exception-values
can be used to allow those tracebacks to be grouped together as well.

Additional changes were to switch to using pytest for running tests.
Part of this is allowing pytest to handle patching stdin, stdout, etc.
and "parametrization" of tests.